+++ src/main.rs
@@ -1,4 +1,5 @@
 use tokio::net::TcpListener;
+use crate::check_changes_job::check_beam_file_changes_job;
 use crate::environment_variables::EnvironmentVariable;
 
 mod routes;
@@ -6,6 +7,7 @@ mod environment_variables;
 mod beam_app;
 mod security;
 mod app_register;
+mod check_changes_job;
 
 #[tokio::main]
 async fn main() {
@@ -20,8 +22,15 @@ async fn main() {
     let address = format!("{}:{}", host, port);
     // Bind to a specific address and port
     let listener = TcpListener::bind(address.clone()).await.unwrap();
-    println!("Server running at http://{}", address );
+    println!("Server running at http://{}", address);
+
+    // Spawn the cron job to run concurrently
+    tokio::spawn(async {
+        // Call the function that runs the cron job
+        check_beam_file_changes_job().await;
+    });
 
     // Use `axum::serve` to run the server
     axum::serve(listener, app).await.unwrap();
+
 }