+++ src/main.rs
@@ -1,6 +1,7 @@
 use tokio::net::TcpListener;
 use crate::check_changes_job::check_beam_file_changes_job;
 use crate::environment_variables::EnvironmentVariable;
+use log::{info, error};
 
 mod routes;
 mod environment_variables;
@@ -11,18 +12,19 @@ mod check_changes_job;
 
 #[tokio::main]
 async fn main() {
+    env_logger::init();
     // Build the application using routes from the `routes` module
     let app = routes::create_router();
     let host = EnvironmentVariable::Host.get_env_var();
     // Ensure no protocol is included in the host
     if host.starts_with("http://") || host.starts_with("https://") {
-        panic!("Host should not include the protocol (e.g., 'http://').");
+        error!("Host should not include the protocol (e.g., 'http://').");
     }
     let port = EnvironmentVariable::Port.get_env_var();
     let address = format!("{}:{}", host, port);
     // Bind to a specific address and port
     let listener = TcpListener::bind(address.clone()).await.unwrap();
-    println!("Server running at http://{}", address);
+    info!("Server running at http://{}", address);
 
     // Spawn the cron job to run concurrently
     tokio::spawn(async {