+++ src/main.rs
@@ -1,5 +1,4 @@
 use tokio::net::TcpListener;
-use crate::check_changes_job::check_beam_file_changes_job;
 use crate::environment_variables::EnvironmentVariable;
 use log::{info, error};
 
@@ -8,8 +7,6 @@ mod environment_variables;
 mod beam_app;
 mod security;
 mod app_register;
-mod check_changes_job;
-mod beam_proxy_reset;
 
 #[tokio::main]
 async fn main() {
@@ -26,13 +23,7 @@ async fn main() {
     // Bind to a specific address and port
     let listener = TcpListener::bind(address.clone()).await.unwrap();
     info!("Server running at http://{}", address);
-
-    // Spawn the cron job to run concurrently
-    tokio::spawn(async {
-        // Call the function that runs the cron job
-        check_beam_file_changes_job().await;
-    });
-
+    
     // Use `axum::serve` to run the server
     axum::serve(listener, app).await.unwrap();