+++ src/config.rs
@@ -5,6 +5,7 @@ use clap::Parser;
 
 #[derive(Debug, Parser)]
 pub struct Config {
+   
     #[clap(long, env, default_value = "0.0.0.0:3000")]
     pub bind_addr: SocketAddr,
 
@@ -16,5 +17,12 @@ pub struct Config {
 
     #[clap(long, env)]
     pub beam_file_path: PathBuf,
+    
+    #[clap(long, env)]
+    pub beam_file_change_check_cron_expression: String,
+    
+    #[clap(long, env)]
+    pub beam_proxy_container_name: String,
+    
 }