+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Authorization ApiKey
 - Register and unregister apps in beam (write env file)
 - Check changes in beam file job
+- Logging with env logger